+++ packages/magento-customer/components/ApolloCustomerError/ApolloCustomerErrorSnackbar.tsx
@@ -1,7 +1,31 @@
 import { ApolloErrorSnackbar, ApolloErrorSnackbarProps } from '@graphcommerce/ecommerce-ui'
+import { Trans } from '@lingui/react'
+import { Button, Link } from '@mui/material'
+import NextLink from 'next/link'
+import { useCustomerSession } from '../../hooks/useCustomerSession'
+import { useAuthorizationErrorMasked } from './useAuthorizationErrorMasked'
 
-type ApolloCustomerErrorSnackbarProps = ApolloErrorSnackbarProps
+export type ApolloCustomerErrorSnackbarProps = ApolloErrorSnackbarProps
 
 export function ApolloCustomerErrorSnackbar(props: ApolloCustomerErrorSnackbarProps) {
-  return <ApolloErrorSnackbar {...props} />
+  const { error, action } = props
+  const [newError, unauthorized] = useAuthorizationErrorMasked(error)
+  const { token } = useCustomerSession()
+
+  return (
+    <ApolloErrorSnackbar
+      {...props}
+      action={
+        unauthorized ? (
+          <NextLink href='/account/signin' passHref>
+            <Button variant='pill' color='secondary'>
+              {token ? <Trans id='Sign in' /> : <Trans id='Create Account' />}
+            </Button>
+          </NextLink>
+        ) : (
+          action
+        )
+      }
+    />
+  )
 }