+++ packages/magento-customer/hooks/useCustomerSession.ts
@@ -1,15 +1,12 @@
 import { useQuery } from '@graphcommerce/graphql'
-import { CustomerTokenDocument } from './CustomerToken.gql'
-import { CustomerTokenFragment } from './CustomerTokenFragment.gql'
+import { CustomerTokenDocument, CustomerTokenQuery } from './CustomerToken.gql'
 
-export type UseCustomerTokenReturn =
-  | (CustomerTokenFragment & {
-      loggedIn: boolean
-      requireAuth: boolean
-    })
-  | (Partial<CustomerTokenFragment> & { loggedIn: false; requireAuth: true })
+type TokenResponse = Omit<NonNullable<CustomerTokenQuery['customerToken']>, '__typename'>
 
-export function useCustomerSession(): UseCustomerTokenReturn {
+export type UseCustomerSessionReturn =
+  | Partial<TokenResponse> & { loggedIn: boolean; requireAuth: boolean }
+
+export function useCustomerSession(): UseCustomerSessionReturn {
   const token = useQuery(CustomerTokenDocument, { ssr: false, fetchPolicy: 'cache-only' }).data
     ?.customerToken