+++ packages/magento-customer/hooks/useFormIsEmailAvailable.tsx
@@ -1,10 +1,9 @@
-import { useQuery } from '@graphcommerce/graphql'
 import { useFormAutoSubmit, useFormGqlQuery, useFormPersist } from '@graphcommerce/react-hook-form'
 import { useEffect, useState } from 'react'
 import { CustomerDocument } from './Customer.gql'
-import { CustomerTokenDocument } from './CustomerToken.gql'
 import { IsEmailAvailableDocument } from './IsEmailAvailable.gql'
 import { useCustomerQuery } from './useCustomerQuery'
+import { useCustomerSession } from './useCustomerSession'
 
 export type UseFormIsEmailAvailableProps = {
   email?: string | null
@@ -13,32 +12,31 @@ export type UseFormIsEmailAvailableProps = {
 
 export function useFormIsEmailAvailable(props: UseFormIsEmailAvailableProps) {
   const { email, onSubmitted } = props
-  const { data: token } = useQuery(CustomerTokenDocument)
+  const { loggedIn, requireAuth } = useCustomerSession()
   const customerQuery = useCustomerQuery(CustomerDocument)
 
   const form = useFormGqlQuery(
     IsEmailAvailableDocument,
     { mode: 'onChange', defaultValues: { email: email ?? '' } },
-    { fetchPolicy: 'cache-and-network' },
+    // { fetchPolicy: 'cache-and-network' },
   )
   const { formState, data, handleSubmit } = form
 
   const submit = handleSubmit(onSubmitted || (() => {}))
-  const autoSubmitting = useFormAutoSubmit({ form, submit })
+  const autoSubmitting = useFormAutoSubmit({ form, submit, forceInitialSubmit: true })
 
   const hasAccount = data?.isEmailAvailable?.is_email_available === false
+
   const { isDirty, isSubmitSuccessful, isSubmitted, isSubmitting, isValid } = formState
 
-  const isLoggedIn = token?.customerToken && token?.customerToken.valid
-
   const [mode, setMode] = useState<'email' | 'signin' | 'signup' | 'signedin' | 'session-expired'>(
-    token?.customerToken && token?.customerToken.valid ? 'signedin' : 'email',
+    loggedIn ? 'signedin' : 'email',
   )
 
   useFormPersist({ form, name: 'IsEmailAvailable' })
 
   useEffect(() => {
-    if (isLoggedIn) {
+    if (loggedIn) {
       setMode('signedin')
       return
     }
@@ -47,19 +45,19 @@ export function useFormIsEmailAvailable(props: UseFormIsEmailAvailableProps) {
     if (!isDirty && isSubmitted && isSubmitSuccessful && isValid)
       setMode(hasAccount ? 'signin' : 'signup')
 
-    if (customerQuery.data?.customer && token && token.customerToken && !token.customerToken.valid)
+    if (customerQuery.data?.customer && requireAuth)
       setMode(isSubmitSuccessful ? 'signin' : 'session-expired')
   }, [
     customerQuery.data?.customer,
     hasAccount,
     isDirty,
-    isLoggedIn,
     isSubmitSuccessful,
     isSubmitted,
     isSubmitting,
     isValid,
-    token,
+    loggedIn,
+    requireAuth,
   ])
 
-  return { mode, form, token, submit, autoSubmitting, hasAccount }
+  return { mode, form, submit, autoSubmitting, hasAccount }
 }