+++ packages/magento-wishlist/components/ProductWishlistChip/ProductWishlistChipBase.tsx
@@ -1,5 +1,9 @@
-import { useQuery, useMutation, useApolloClient } from '@graphcommerce/graphql'
-import { CustomerTokenDocument } from '@graphcommerce/magento-customer'
+import { useMutation, useApolloClient } from '@graphcommerce/graphql'
+import {
+  useCustomerQuery,
+  useCustomerSession,
+  useGuestQuery,
+} from '@graphcommerce/magento-customer'
 import {
   IconSvg,
   iconHeart,
@@ -10,7 +14,7 @@ import {
 } from '@graphcommerce/next-ui'
 import { i18n } from '@lingui/core'
 import { Trans } from '@lingui/react'
-import { SxProps, Theme, IconButton, Box } from '@mui/material'
+import { SxProps, Theme, IconButton, Box, IconButtonProps } from '@mui/material'
 import PageLink from 'next/link'
 import { useState, useEffect } from 'react'
 import { useWishlistEnabled } from '../../hooks'
@@ -27,6 +31,7 @@ const ignoreProductWishlistStatus =
 export type ProductWishlistChipProps = ProductWishlistChipFragment & { sx?: SxProps<Theme> } & {
   selectedOptions?: string[]
   showFeedbackMessage?: boolean
+  buttonProps?: IconButtonProps
 }
 
 const compName = 'ProductWishlistChipBase' as const
@@ -34,24 +39,18 @@ const parts = ['root', 'wishlistIcon', 'wishlistIconActive', 'wishlistButton'] a
 const { classes } = extendableComponent(compName, parts)
 
 export function ProductWishlistChipBase(props: ProductWishlistChipProps) {
-  const { name, sku, showFeedbackMessage, selectedOptions = [], sx = [] } = props
+  const { name, sku, showFeedbackMessage, selectedOptions = [], buttonProps, sx = [] } = props
 
   const [inWishlist, setInWishlist] = useState(false)
   const [displayMessageBar, setDisplayMessageBar] = useState(false)
 
-  const { data: token } = useQuery(CustomerTokenDocument)
+  const { loggedIn } = useCustomerSession()
   const [addWishlistItem] = useMutation(AddProductToWishlistDocument)
   const [removeWishlistItem] = useMutation(RemoveProductFromWishlistDocument)
-  const isLoggedIn = token?.customerToken && token?.customerToken.valid
 
-  const { data: GetCustomerWishlistData, loading } = useQuery(GetIsInWishlistsDocument, {
-    skip: !isLoggedIn,
-  })
+  const { data: GetCustomerWishlistData, loading } = useCustomerQuery(GetIsInWishlistsDocument)
 
-  const { data: guestWishlistData } = useQuery(GuestWishlistDocument, {
-    ssr: false,
-    skip: isLoggedIn === true,
-  })
+  const { data: guestWishlistData } = useGuestQuery(GuestWishlistDocument)
 
   const { cache } = useApolloClient()
 
@@ -80,7 +79,7 @@ export function ProductWishlistChipBase(props: ProductWishlistChipProps) {
 
   useEffect(() => {
     // Do not display wishlist UI to guests when configured as customer only
-    if (hideForGuest && !isLoggedIn) {
+    if (hideForGuest && !loggedIn) {
       return
     }
 
@@ -89,17 +88,17 @@ export function ProductWishlistChipBase(props: ProductWishlistChipProps) {
     }
 
     // Mark as active when product is available in either customer or guest wishlist
-    if (isLoggedIn && !loading) {
+    if (loggedIn && !loading) {
       const inWishlistTest =
         GetCustomerWishlistData?.customer?.wishlists[0]?.items_v2?.items.map(
           (item) => item?.product?.sku,
         ) || []
       setInWishlist(inWishlistTest.includes(sku))
-    } else if (!isLoggedIn) {
+    } else if (!loggedIn) {
       const inWishlistTest = guestWishlistData?.guestWishlist?.items.map((item) => item?.sku) || []
       setInWishlist(inWishlistTest.includes(sku))
     }
-  }, [isLoggedIn, sku, loading, GetCustomerWishlistData, guestWishlistData])
+  }, [loggedIn, sku, loading, GetCustomerWishlistData, guestWishlistData])
 
   const preventAnimationBubble: React.MouseEventHandler<HTMLButtonElement> = (e) => {
     e.preventDefault()
@@ -113,7 +112,7 @@ export function ProductWishlistChipBase(props: ProductWishlistChipProps) {
       return
     }
 
-    if (isLoggedIn) {
+    if (loggedIn) {
       if (inWishlist && !ignoreProductWishlistStatus) {
         const wishlistItemsInSession =
           GetCustomerWishlistData?.customer?.wishlists[0]?.items_v2?.items || []
@@ -172,6 +171,7 @@ export function ProductWishlistChipBase(props: ProductWishlistChipProps) {
         onMouseDown={preventAnimationBubble}
         size='small'
         className={classes.wishlistButton}
+        {...buttonProps}
         sx={[
           (theme) => ({
             padding: theme.spacings.xxs,
@@ -218,7 +218,7 @@ export function ProductWishlistChipBase(props: ProductWishlistChipProps) {
     </Box>
   )
 
-  return !hideForGuest || isLoggedIn ? output : null
+  return !hideForGuest || loggedIn ? output : null
 }
 
 ProductWishlistChipBase.defaultProps = {