+++ packages/next-ui/CHANGELOG.md
@@ -1,5 +1,37 @@
 # Change Log
 
+## 4.8.4
+
+### Patch Changes
+
+- [#1509](https://github.com/graphcommerce-org/graphcommerce/pull/1509) [`0ab7c5465`](https://github.com/graphcommerce-org/graphcommerce/commit/0ab7c5465441cba9bf8cd185a6790ce2f443f4ed) Thanks [@paales](https://github.com/paales)! - SidebarGallery improvements (product page):
+
+  - Prevent vertical scrolling
+  - Disable zoom fab when there are no images
+  - Hide scroller dots when there in only one image
+  - Make sure the prev/next buttons are shown as expected
+
+- Updated dependencies [[`0ab7c5465`](https://github.com/graphcommerce-org/graphcommerce/commit/0ab7c5465441cba9bf8cd185a6790ce2f443f4ed)]:
+  - @graphcommerce/framer-scroller@2.1.15
+
+## 4.8.3
+
+### Patch Changes
+
+- [#1487](https://github.com/graphcommerce-org/graphcommerce/pull/1487) [`afc67103d`](https://github.com/graphcommerce-org/graphcommerce/commit/afc67103d0e00583e274465036fd287537f95e79) Thanks [@paales](https://github.com/paales)! - When additing an additional breakpoint it would throw a typescript error
+
+- Updated dependencies []:
+  - @graphcommerce/framer-scroller@2.1.14
+
+## 4.8.2
+
+### Patch Changes
+
+- [#1485](https://github.com/graphcommerce-org/graphcommerce/pull/1485) [`c8c246b8a`](https://github.com/graphcommerce-org/graphcommerce/commit/c8c246b8aaab0621b68a2fca2a1c529a56fad962) Thanks [@paales](https://github.com/paales)! - TextInputNumber: when adding a label it should be displayed properly
+
+- Updated dependencies []:
+  - @graphcommerce/framer-scroller@2.1.13
+
 ## 4.8.1
 
 ### Patch Changes